[Silver III] Title: 주유소, Time: 356 ms, Memory: 30468 KB -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -61,8 +61,8 @@ public class Main {
 			arr[1][i]=Integer.parseInt(st.nextToken());
 		}
 		
-		long dist=0;
-		long price=0;
+		int dist=0;
+		int price=0;
 		int idx=0;
 		
 		for(int i=0;i<N-1;i++) {
